Arbitration UI updates. Relative finalization times in UI. Add Funds to accepted offer.

+++ contracts/IMarketplace.sol
@@ -52,7 +52,7 @@ interface IMarketplace {
   function makeOffer(
     uint listingID,
     bytes32 _ipfsHash,   // IPFS hash containing offer data
-    uint32 _finalizes,   // Timestamp an accepted offer will finalize
+    uint256 _finalizes,   // Timestamp an accepted offer will finalize
     address _affiliate,  // Address to send any required commission to
     uint256 _commission, // Amount of commission to send in Origin Token if offer finalizes
     uint _value,         // Offer amount in ERC20 or Eth
